backend/requirements.txt
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.35
pandas>=2.2.2
pyarrow>=17.0.0
python-dotenv>=1.0.1
pydantic>=2.8.2
tabulate>=0.9.0
fastapi>=0.115.5
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
boto3>=1.34.0
